requirements.txt
greenlet==3.0.1
mysql-connector-python==8.2.0
numpy==1.26.2
pandas==2.1.3
protobuf==4.21.12
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
